pyproject.toml
[tool.poetry]
name = "lekzia-13-01"
version = "0.1.0"
description = ""
authors = []


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
alembic = "^1.13.0"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
uvicorn = "^0.24.0.post1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
pydantic = { extras = ["email"], version = "^2.5.2" }
python-multipart = "^0.0.6"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.1.1"
redis = "^6.4.0"
fastapi-limiter = "^0.1.6"
bcrypt = "<4.0"
jinja2 = "^3.1.6"
cloudinary = "^1.44.1"
pytest = "^8.4.2"


[tool.poetry.group.dev.dependencies]
sphinx = "7.2.6"
pytest-asyncio = "^1.2.0"
coverage = "^7.11.0"


[tool.poetry.group.test.dependencies]
aiosqlite = "^0.21.0"
pytest-asyncio = "^1.2.0"
httpx = "0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = ["tests"]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"
